additional docs and cleanup
