+++ Sources/Vapor/Utilities/Exports.swift
@@ -3,6 +3,8 @@
 @_exported import Async
 @_exported import CodableKit
 @_exported import Core
+@_exported import Console
+@_exported import Command
 @_exported import DatabaseKit
 @_exported import FormURLEncoded
 @_exported import HTTP
@@ -13,7 +15,4 @@
 @_exported import Routing
 @_exported import Service
 @_exported import TemplateKit
-@_exported import class Foundation.URLSession
-@_exported import struct Foundation.Data
-@_exported import struct Foundation.Date
 @_exported import WebSocket
